requirements.txt
Babel==2.15.0
beautifulsoup4==4.12.3
certifi==2024.6.2
charset-normalizer==3.3.2
idna==3.7
pillow==10.3.0
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.16.0
soupsieve==2.5
tk==0.1.0
tkcalendar==1.6.1
urllib3==2.2.2
